requirements.txt
streamlit>=1.28.0
openai>=1.3.0
elevenlabs>=0.2.26
azure-cognitiveservices-speech>=1.32.0
moviepy>=1.0.3
opencv-python>=4.8.0
pillow>=10.0.0
pydub>=0.25.1
openai-whisper>=20231117
ffmpeg-python>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
librosa>=0.10.0
soundfile>=0.12.1
torch>=2.0.0
torchaudio>=2.0.0
pyaudio>=0.2.11 